Split game components from main bundle
