+++ package.json
@@ -44,6 +44,7 @@
     "react-feather": "^1.1.0",
     "react-hint": "^3.1.0",
     "react-intl": "^2.4.0",
+    "react-loadable": "^5.4.0",
     "react-markdown": "^3.3.4",
     "react-modal": "^3.4.5",
     "react-redux": "^5.0.7",
