+++ src/components/empty-overview/style.scss
@@ -2,7 +2,7 @@
 
 .EmptyOverview {
   background-color: $dark3;
-  padding: 1.5rem;
+  position: relative;
 }
 
 .EmptyOverview-Header1 {
