+++ src/components/sector-map/loading/index.js
@@ -1,15 +1,11 @@
-import React from 'react';
-import { FormattedMessage } from 'react-intl';
+import React, { Fragment } from 'react';
 
-import ContentContainer from 'primitives/container/content-container';
-import AbsoluteContainer from 'primitives/container/absolute-container';
-import Header, { HeaderType } from 'primitives/text/header';
-import Spinner from 'primitives/other/spinner';
+import Loading from 'primitives/other/loading';
 import HexMap from 'components/hex-map';
 
 import hexGenerator from 'utils/hex/generator';
 
-export default function Loading() {
+export default function SectorLoading() {
   const { hexes } = hexGenerator({
     renderSector: false,
     height: window.innerHeight,
@@ -17,20 +13,13 @@ export default function Loading() {
   });
 
   return (
-    <div>
+    <Fragment>
       <HexMap
         height={window.innerHeight}
         width={window.innerWidth}
         hexes={hexes}
       />
-      <AbsoluteContainer>
-        <ContentContainer direction="column" align="center" justify="center">
-          <Spinner size={100} />
-          <Header type={HeaderType.header2}>
-            <FormattedMessage id="misc.loadingSector" />
-          </Header>
-        </ContentContainer>
-      </AbsoluteContainer>
-    </div>
+      <Loading />
+    </Fragment>
   );
 }
