+++ src/primitives/form/dropdown/style.scss
@@ -1,5 +1,4 @@
 @import 'styles/theme';
-@import 'react-select/dist/react-select';
 
 .Dropdown {
   font-size: 11px;
