+++ src/styles/global.scss
@@ -1,4 +1,5 @@
-@import 'react-redux-toastr/src/styles/index.scss';
+@import 'react-redux-toastr/src/styles/index';
+@import 'react-select/scss/default';
 @import 'styles/theme';
 
 body {
